make python example work
Hey there, completely new to `openCV` in general and to the `bioinspired` module in particular. I wanted to give this module a try with python and found that the example source code doesn't work with current openCV (that's 3.4.1 on my Fedora machine).  Apparently, there was a change in how a `Retina` object is created. I fixed the python source code correspondingly. I also changed the C++ source code in this doc file, although I think this needs further updating as there are more changes in the actual source code.