sfm module calls imread function, which is in imgcodecs module
This built correctly before because tests were run with the highgui module enabled.
The sfm module links against the features2d module, which brings in the highgui module as an optional target if enabled.
Through this path, imread was working correctly despite the file including the wrong header and not declaring its linking dependency.